fix(appstore): fix import cycle
